Shared ViewfinderGestureLayer: full gesture parity on the director
Extract the 1:1 monitor's viewfinder touch surface — tap-to-focus
(FocusPointMapping + reticle), double-tap camera flip, pinch zoom — into
ViewfinderGestureLayer, owned by neither screen. The monitor uses it
unchanged (snapshot tests pin the pixels); the director drops in the
same layer with its callbacks routed to the focused camera, gaining
pinch zoom and double-tap flip.

This also fixes the director's reticle offset: the layer, the live
frame, and the measured size now share one ZStack under a single
ignoresSafeArea, so the reticle draws under the finger instead of a
safe-area inset below it.

The lane snapshot now projects supportsFocusPoint, and the layer
withholds the focus tap entirely for a camera that can't focus at a
point — no reticle promising nothing, and no paywall for a feature the
focused camera doesn't have. Grid mode is unchanged: a tile tap selects
the camera; precision gestures live in the full-size view.
