+++ RemoteCam/MonitorView.swift
@@ -12,7 +12,6 @@ import Combine
 /// whichever edge the screen's *shape* makes cheap — see `MonitorChromeLayout`.
 struct MonitorView: View {
     @ObservedObject var viewModel: MonitorViewModel
-    @State private var zoomAtGestureStart: CGFloat?
     /// Peer-link state; the reconnect overlay is a function of it.
     @ObservedObject var peerLink: PeerLinkStatus = .shared
 
@@ -38,11 +37,6 @@ struct MonitorView: View {
     /// Toggles the connected camera's local-preview mode (on ⇄ standby).
     let onToggleCameraStandby: () -> Void
 
-    /// The live focus reticle (view-space position + identity to re-trigger the
-    /// animation on each tap). Local UI only — no round-trip to the camera.
-    @State private var focusReticle: FocusReticle?
-    /// The preview area's measured size, for tap → normalized-image mapping.
-    @State private var previewSize: CGSize = .zero
     @State private var isTrayOpen = false
 
     var body: some View {
@@ -75,7 +69,6 @@ struct MonitorView: View {
         // already draw their own shape. Not .plain — that also drops the
         // style's hit region, leaving material fills unclickable.
         .buttonStyle(.borderless)
-        .onPreferenceChange(PreviewSizePreferenceKey.self) { previewSize = $0 }
         .statusBarHidden()
     }
 
@@ -123,61 +116,21 @@ struct MonitorView: View {
                 .opacity(viewModel.isPreviewStale ? 0.65 : 1)
                 .animation(.easeInOut(duration: 0.25), value: viewModel.isPreviewStale)
 
-            // Preview gestures sit BELOW the interactive chrome, so a tap on a
-            // control is never stolen as a focus tap. Double tap toggles the
-            // camera; a single tap focuses (simultaneous so the reticle is
-            // instant — an exclusive gesture would stall it for the double-tap
-            // window); pinch zooms. The translation guard keeps drags and
-            // pinches from focusing.
-            Color.clear
-                .contentShape(Rectangle())
-                .onTapGesture(count: 2) {
-                    onToggleCamera()
-                }
-                .simultaneousGesture(
-                    DragGesture(minimumDistance: 0)
-                        .onEnded { value in
-                            if abs(value.translation.width) < 10, abs(value.translation.height) < 10 {
-                                handleFocusTap(at: value.location)
-                            }
-                        }
-                )
-                .simultaneousGesture(
-                    MagnificationGesture()
-                        .onChanged { value in
-                            if zoomAtGestureStart == nil {
-                                zoomAtGestureStart = viewModel.currentZoomFactor
-                            }
-                            let start = zoomAtGestureStart!
-                            onZoomChange(viewModel.zoomScale.pinched(from: start, magnification: value))
-                        }
-                        .onEnded { _ in
-                            zoomAtGestureStart = nil
-                        }
-                )
-
-            focusReticleOverlay
+            // Gestures (focus tap, double-tap flip, pinch zoom) + reticle, in
+            // the preview's own coordinate space, BELOW the interactive chrome
+            // so a tap on a control is never stolen as a focus tap.
+            ViewfinderGestureLayer(
+                cameraImage: { viewModel.frames.cameraImage },
+                zoomScale: viewModel.zoomScale,
+                currentZoomFactor: viewModel.currentZoomFactor,
+                focusEnabled: true,
+                onFocusTap: onFocusTap,
+                onDoubleTap: onToggleCamera,
+                onZoomChange: onZoomChange)
 
             countdownOverlay
         }
-        .background(
-            GeometryReader { geo in
-                Color.clear.preference(key: PreviewSizePreferenceKey.self, value: geo.size)
-            }
-        )
         .ignoresSafeArea(edges: Self.previewBleedEdges)
-    }
-
-    /// The focus reticle, drawn in the preview's coordinate space. Non-interactive
-    /// so it never eats a subsequent tap.
-    @ViewBuilder
-    private var focusReticleOverlay: some View {
-        if let reticle = focusReticle {
-            FocusReticleView()
-                .id(reticle.id)
-                .position(reticle.point)
-                .allowsHitTesting(false)
-        }
     }
 
     /// The self-timer, centered and large. The capture happens across the room:
@@ -198,25 +151,6 @@ struct MonitorView: View {
                 .id(viewModel.timerValue)
                 .transition(.scale(scale: 1.15).combined(with: .opacity))
                 .allowsHitTesting(false)
-        }
-    }
-
-    /// Maps a preview tap into a normalized image point and, if it landed on the
-    /// image (not the letterbox), shows the reticle and forwards it to the camera.
-    private func handleFocusTap(at location: CGPoint) {
-        guard let image = viewModel.frames.cameraImage,
-              let normalized = FocusPointMapping.normalizedImagePoint(
-                tap: location, viewSize: previewSize, imageSize: image.size)
-        else { return }
-        showFocusReticle(at: location)
-        onFocusTap(normalized)
-    }
-
-    private func showFocusReticle(at point: CGPoint) {
-        let reticle = FocusReticle(point: point)
-        focusReticle = reticle
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.6) {
-            if focusReticle?.id == reticle.id { focusReticle = nil }
         }
     }
 
@@ -883,49 +817,6 @@ struct LiveFrameView: View {
                         .foregroundColor(.white.opacity(0.5))
                 )
         }
-    }
-}
-
-// MARK: - Tap to Focus
-
-/// One tap-to-focus reticle: its view-space position plus an identity so a new
-/// tap re-instantiates `FocusReticleView` and replays its animation.
-struct FocusReticle: Equatable {
-    let id = UUID()
-    let point: CGPoint
-}
-
-/// Measures the preview area's size so a tap can be mapped to a normalized image
-/// point. A preference key avoids mutating `@State` during layout.
-private struct PreviewSizePreferenceKey: PreferenceKey {
-    static var defaultValue: CGSize = .zero
-    static func reduce(value: inout CGSize, nextValue: () -> CGSize) {
-        let next = nextValue()
-        if next != .zero { value = next }
-    }
-}
-
-/// The animated focus square: a yellow reticle that pulses in and fades. Purely
-/// local tap feedback — the focus command travels separately.
-struct FocusReticleView: View {
-    @State private var scale: CGFloat = 1.25
-    @State private var opacity: Double = 0
-
-    var body: some View {
-        RoundedRectangle(cornerRadius: 6)
-            .stroke(Color.yellow, lineWidth: 1.5)
-            .frame(width: 78, height: 78)
-            .scaleEffect(scale)
-            .opacity(opacity)
-            .onAppear {
-                withAnimation(.easeOut(duration: 0.12)) {
-                    scale = 1.0
-                    opacity = 1
-                }
-                withAnimation(.easeIn(duration: 0.2).delay(0.35)) {
-                    opacity = 0
-                }
-            }
     }
 }
